requirements.txt
python-fasthtml
pydantic
openai
python-dotenv